Add musician matching UI with foriio-style design
Features:
- Implement musician listing page with grid layout
- Add musician detail page showing tracks with BPM/key/genre
- Move existing audio analysis to /upload route
- Create reusable MusicianCard component
- Add component tests using React Testing Library
- Fix TypeScript lint warnings (unused variables)

Design:
- foriio-inspired clean UI with rounded cards
- Tailwind CSS for consistent styling
- No emojis, professional appearance
- Responsive grid layout (1/2/3 columns)

Testing:
- Add MusicianCard component test
- Add HomePage rendering test
